move new changes back into file to maintain history
